Added setup_cg and teardown_cg routines to copy g, register u,w.
g doesn't change across cg iterations, so it can be copied once
and saved.

u,w aren't huge (for now), so it makes sense to keep them registered
for the whole cg step.

Also, some BLOCK/STREAM-related debugging (jp*NUM_BLOCK_MAX)

At this point, if there are 4 streams or so, then everything overlaps
nicely.  There is room to make it faster, but less than 1.5 seconds
are spent on the GPU, out of a 20 second run, so its time to expand
the scope of the GPU port.