Moved back to poly-order 32 elements, final touches on ax.
32x32x32 elements are slightly more realistic.  I tried
increasing the number of elements (all the way up to the maximum
heap size) to increase the amount of time spent in ax, without much
success.  The gpu portion of the code takes less than 10% of the
runtime, so not much else can be done without porting the rest of cg.