Brought this back up to nekbone-3.0
The preconditioner is still just a copy, though renamed, so no
change there.

The only real change to driver was putting a bootstrap call to
maskit before cg_cuda_init.  That mask is then copied to the gpu
and used to properly implement maskit.

This passes both single and multi-core.